+++ docs/NEWS.md
@@ -1,6 +1,7 @@
 ### UNRELEASED
 
  * *feature:* allow decimal values such as "1.5G" with "**--size**", "**--rate-limit**", "**--buffer-size**", and "**error-skip-block**" ([#35](https://codeberg.org/a-j-wood/pv/issues/35))
+ * *feature:* new bar formats "**%{bar-default}**", "**%{bar-block}**", "**%{bar-granular}**", "**%{bar-shaded}**"
  * *docs:* simplified the synopsis section of the manual
 
 ### 1.9.15 - 8 December 2024